+++ matrix_mult.h
@@ -1,5 +1,5 @@
 //matrix multiplication 
-int ** matrix_mult(int **A, int** B){
+double ** matrix_mult(double **A, double** B){
     int m = sizeof(A)/sizeof(A[0]);
     int n = sizeof(A[0])/sizeof(A[0][0]);
     int n_b = sizeof(B)/sizeof(B[0]);
@@ -7,7 +7,7 @@ int ** matrix_mult(int **A, int** B){
     if(n != n_b){
         return nullptr;
     }
-    int **C;
+    double **C;
     for(int i=0; i<m; i++){
         for(int j=0; j<p; j++){
             C[i][j]=0;
